requirements.txt
cffi==1.17.1
curl_cffi==0.10.0
python-dotenv==1.1.0
